--- a/nntrainer/tensor/cpu_backend/arm/arm_compute_backend.h
+++ b/nntrainer/tensor/cpu_backend/arm/arm_compute_backend.h
@@ -490,7 +490,7 @@ void nntr_qsi4cxp_qs4cxs1s0_rhs_pack(size_t n, size_t k,
  * regards the weight to be transpoed.
  */
 template <typename T = float>
-void nntr_gemm_qai8dxp_qsi4cxp_unpacked(
+uint32_t nntr_gemm_qai8dxp_qsi4cxp_unpacked(
   size_t m, size_t n, size_t k, void *lhs_native_mtx,
   void *rhs_native_mtx_qs4cx, void *rhs_scales, T *dst_mtx, bool transB = true,
   T lower_bound = std::numeric_limits<T>::lowest(),