--- a/nntrainer/tensor/cpu_backend/arm/arm_compute_backend_fp16.cpp
+++ b/nntrainer/tensor/cpu_backend/arm/arm_compute_backend_fp16.cpp
@@ -467,13 +467,13 @@ void nntr_quant_qs4cx_f32(size_t n, size_t k, void *rhs_native_mtx_f32,
 }
 
 template <>
-void nntr_gemm_qai8dxp_qsi4cxp_unpacked(size_t m, size_t n, size_t k,
+uint32_t nntr_gemm_qai8dxp_qsi4cxp_unpacked(size_t m, size_t n, size_t k,
                                    void *lhs_native_mtx_f32,
                                    void *rhs_native_mtx_qs4cx,
                                    void *rhs_scales_f32, float *dst_mtx_f32,
                                    bool transB, float lower_bound,
                                    float upper_bound) {
-  nntr_kai_gemm_qai8dxp_qsi4cxp_rtp(
+  return nntr_kai_gemm_qai8dxp_qsi4cxp_rtp(
     m, n, k, lhs_native_mtx_f32, rhs_native_mtx_qs4cx, rhs_scales_f32,
     dst_mtx_f32, transB, lower_bound, upper_bound);
 }