--- a/test/jni/meson.build
+++ b/test/jni/meson.build
@@ -6,7 +6,7 @@ if get_option('debug')
 endif
 
 if get_option('enable-fp16')
-   flags += '-march=armv8.2-a+fp16+dotprod'
+   flags += '-march=armv8.2-a+fp16+dotprod+i8mm'
 endif
 
 flags += '-O@0@'.format(get_option('optimization'))