… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp4x8_4x4x32_neon_i8mm.h
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp4x8_8x4x32_neon_i8mm.c
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp4x8_8x4x32_neon_i8mm.h
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp8x8_4x8x32_neon_i8mm.c
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp8x8_4x8x32_neon_i8mm.h
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp8x8_8x8x32_neon_i8mm.c
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/kai_matmul_clamp_f32_qai8dxp4x8_qsi4cxp8x8_8x8x32_neon_i8mm.h
… nntrainer/tensor/cpu_backend/arm/kai/matmul_clamp_f32_qai8dxp_qsi4cxp/meson.build
… nntrainer/tensor/cpu_backend/arm/neon_kleidiai.cpp
… nntrainer/tensor/cpu_backend/cpu_backend.h
… nntrainer/tensor/cpu_backend/fallback/fallback.cpp
… nntrainer/tensor/cpu_backend/fallback/fallback.h
… nntrainer/tensor/cpu_backend/fallback/fallback_fp16.cpp
… nntrainer/tensor/cpu_backend/fallback/fallback_internal.cpp
… nntrainer/tensor/cpu_backend/fallback/fallback_internal.h
… nntrainer/tensor/cpu_backend/fallback/fallback_internal_fp16.cpp
… nntrainer/tensor/cpu_backend/x86/x86_compute_backend.cpp
… nntrainer/tensor/cpu_backend/x86/x86_compute_backend.h
… nntrainer/tensor/cpu_backend/x86/x86_compute_backend_fp16.cpp
… packaging/nntrainer.spec
… test/jni/Android.mk
… test/unittest/jni/Android.mk
… test/unittest/unittest_nntrainer_cpu_backend_fp16.cpp